+++ src/event.rs
@@ -8,7 +8,7 @@ use std::time::{Instant};
 use serde_json as json;
 use parking_lot::MutexGuard;
 use glutin::{self, ModifiersState, Event, ElementState};
-use copypasta::{Clipboard, Load, Store};
+use copypasta::{Clipboard, Load, Store, Buffer as ClipboardBuffer};
 
 use ansi::{Handler, ClearMode};
 use grid::Scroll;
@@ -64,7 +64,7 @@ impl<'a, N: Notify + 'a> input::ActionContext for ActionContext<'a, N> {
         self.terminal.clear_screen(ClearMode::Saved);
     }
 
-    fn copy_selection(&self, buffer: ::copypasta::Buffer) {
+    fn copy_selection(&self, buffer: ClipboardBuffer) {
         if let Some(selected) = self.terminal.selection_to_string() {
             if !selected.is_empty() {
                 Clipboard::new()
@@ -238,6 +238,7 @@ pub struct Processor<N> {
     last_modifiers: ModifiersState,
     pending_events: Vec<Event>,
     window_changes: WindowChanges,
+    save_to_clipboard: bool,
 }
 
 /// Notify that the terminal was resized
@@ -281,6 +282,7 @@ impl<N: Notify> Processor<N> {
             last_modifiers: Default::default(),
             pending_events: Vec::with_capacity(4),
             window_changes: Default::default(),
+            save_to_clipboard: config.selection().save_to_clipboard,
         }
     }
 
@@ -442,6 +444,7 @@ impl<N: Notify> Processor<N> {
                 mouse_config: &self.mouse_config,
                 key_bindings: &self.key_bindings[..],
                 mouse_bindings: &self.mouse_bindings[..],
+                save_to_clipboard: self.save_to_clipboard,
             };
 
             let mut window_is_focused = window.is_focused;
@@ -496,5 +499,6 @@ impl<N: Notify> Processor<N> {
         self.key_bindings = config.key_bindings().to_vec();
         self.mouse_bindings = config.mouse_bindings().to_vec();
         self.mouse_config = config.mouse().to_owned();
+        self.save_to_clipboard = config.selection().save_to_clipboard;
     }
 }