+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
         viewport to extend to the top of the window.
     - `buttonless` - Similar to transparent but also removed the buttons.
 - Add support for changing the colors from 16 to 256 in the `indexed_colors` config section
+- New config `save_to_clipboard: true` in `selection` section: copy selected
+  string to primary clipboard
 
 ### Changed